Sonar code smells and test coverage
