+++ packages/core/src/index.ts
@@ -4,5 +4,6 @@ export * from './fhirpath';
 export * from './format';
 export * from './outcomes';
 export * from './search';
+export * from './searchparams';
 export * from './types';
 export * from './utils';
